backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["SotaChambers <sota.streamlit@gmail.com>"]
packages = [
    { include = "app" },
]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.114.2"
uvicorn = "^0.30.6"
pydantic = "^2.9.1"
alembic = "^1.13.2"
sqlalchemy = "^2.0.35"
python-dotenv = "^1.0.1"
pydantic-settings = "^2.5.2"
pymysql = "^1.1.1"
cryptography = "^43.0.1"
pandas = "^2.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
